Update ytDownloader.py
included if clause

+++ Python Projekte/Youtube Downloader/ytDownloader.py
@@ -11,7 +11,7 @@ yt = YouTube(link)
 #print('Title: ', yt.title)
 #print('Views: ', yt.views)
 
-yd = yt.streams.get_highest_resolution() #get highest resolution lol
+yd = yt.streams.get_highest_resolution()
 
 #save (safe) and sound
 saved_video = yd.download() #note: output saved to the current working directory
@@ -21,4 +21,13 @@ video = VideoFileClip(saved_video)
 video.audio.write_audiofile(saved_video.replace(".mp4", ".mp3")) 
 
 # to delete the .mp4 video file:
-os.remove(saved_video)
+answer = input("Möchtest du das Video noch löschen? [y/n] ")
+yes = {'yes', 'y', 'ye', ''}
+no = {'no', 'n', 'noo'}
+# you rly want to delete the video?
+if answer in yes:
+    os.remove(saved_video)
+elif answer in no:
+    pass
+else:
+    sys.stdout.write("Bitte antworte mit 'yes' or 'no'")
